Seperate controllers from the routes file.
Now routes file will only handle routes.
all of the complex logic will be handled by controllers.
